+++ include/ydsh/ydsh.h
@@ -314,6 +314,13 @@ const char *DSState_copyright();
 
 unsigned int DSState_featureBit();
 
+/**
+ * get full path of current executable path (may be real path of argv[0])
+ * @return
+ * if cannot resolve path, return null.
+ * after call it, manually call free() to release buffer
+ */
+char *DSState_getExecutablePath();
 
 /* for input completion */