+++ src/ydsh.cpp
@@ -737,6 +737,47 @@ unsigned int DSState_featureBit() {
     return flag;
 }
 
+#ifdef __APPLE__
+#include <mach-o/dyld.h>
+
+char *DSState_getExecutablePath() {
+    uint32_t bufSize = 0;
+    _NSGetExecutablePath(nullptr, &bufSize);    // get buffer size
+    char *buf = static_cast<char*>(malloc(sizeof(char) * (bufSize + 1)));
+    char *real = nullptr;
+    if(_NSGetExecutablePath(buf, &bufSize) == 0) {
+        real = realpath(buf, nullptr);
+    }
+    free(buf);
+    return real;
+}
+
+#elif __linux__
+char *DSState_getExecutablePath() {
+    const char *selfexec = "/proc/self/exec";
+    struct stat st;
+    if(lstat(selfexec, &st) == -1) {
+        return nullptr;
+    }
+    char *link = static_cast<char*>(malloc(sizeof(char) * (st.st_size + 1)));
+    if(!link) {
+        return nullptr;
+    }
+    ssize_t len = readlink(selfexec, link, sizeof(char) * (st.st_size + 1));
+    if(len < 0 || len > st.st_size) {
+        free(link);
+        return nullptr;
+    }
+    link[len] = '\0';
+    return link;
+}
+
+#else
+char *DSState_getExecutablePath() {
+    return nullptr;
+}
+#endif
+
 unsigned int DSState_complete(DSState *st, DSCompletionOp op, unsigned int index, const char **value) {
     GUARD_NULL(st, 0);